refactor(workstation): extract detail / inspector / preview family from inkRuntime (#890 phase 5a.6)
Sixth batch of surface extractions, building on #901 (5a.5). The detail
family is 11 functions forming a cohesive group — they all render the
right-side detail pane content. Each is dispatched to by
renderDetailPanel (which stays in inkRuntime as the thin dispatcher to
both surface renderers and overlay renderers).

7 exported renderers moved to src/workstation/surfaces/detail/index.ts:
  - renderHistoryInspector       (the main commit inspector, 174 LOC)
  - renderCommitDiffDetail       (commit-sourced diff right panel)
  - renderComposeContextPanel    (compose / pending-commit right panel)
  - renderBranchPreviewPanel     (P4.1 — fzf-style branch preview)
  - renderTagPreviewPanel        (P4.1 — fzf-style tag preview)
  - renderStashPreviewPanel      (P4.1 — fzf-style stash preview)
  - renderCommitPanel            (status compose right panel)

4 private helpers moved with the family (only the surface uses them):
  - renderInspectorActionsSection  (per-entity action cheat-sheet)
  - renderInspectorRefs            (OSC 8 ref-list wrapper)
  - renderCommitFileList           (shared by inspector + commit diff)
  - renderPreviewPanel             (shared by 3 preview panels)

Detail surface file: 703 LOC. Same pattern as the history surface in 5a.5
— only the top-level dispatchable functions are exported; private helpers
stay file-local.

inkRuntime.ts: 4,040 -> 3,391 LOC (-649 this PR; -2,648 cumulative since
phase 5a started, -44% off the original 6,039). 16 imports pruned that
became unused post-extraction:
  formatHyperlink (chrome/hyperlinks — whole import gone)
  PreviewLine, formatBranchPreview, formatStashPreview, formatTagPreview
    (chrome/previewPane — whole import gone)
  formatLogInkLoading (chrome/surfaceStates — whole import gone)
  wrapCells (chrome/text)
  ProviderRepository (git/providerData)
  InspectorAction, InspectorActionContext, getInspectorActions
    (chrome/inspectorActions — whole import gone)
  buildCommitUrl, buildRefUrl, compactHash, formatChangedFileStats,
    statusCodeColor (runtime/utils)

Three more chrome/ module imports went away entirely
(chrome/hyperlinks, chrome/previewPane, chrome/surfaceStates,
chrome/inspectorActions) — the orchestration layer is now nearly free of
chrome dependencies; the surface modules consume chrome directly. Exactly
the dependency direction the workstation architecture is supposed to
enforce.

What's left in inkRuntime.ts at 3,391 LOC:
  - LogInkApp React component (~1,940 LOC, phase 5b target)
  - renderDetailPanel dispatcher (~87 LOC, stays — it dispatches between
    extracted surfaces AND overlays still in this file)
  - Chrome render helpers (header, sidebar, footer, ~1,200 LOC, phase 5a.7)
  - Overlays + input prompt + confirmation + palette + help (phase 5a.7)
  - Imports + small orchestration + startInkInteractiveLog + loadInkRuntime

Validation: tsc clean, lint clean, 680/680 suites & 6571/6571 tests
passing (same baseline as post-5a.5). No behavior change — every
detail-pane variant renders byte-identically before/after.
